+++ ParallelCalc/Program.cs
@@ -28,7 +28,7 @@ internal class Program
     static void Main(string[] args)
     {
         int numberOfThreads = Environment.ProcessorCount;
-        IRandomGenerator<int> randomGenerator = new RangomIntGenerator();
+        IRandomGenerator<long> randomGenerator = new RangomIntGenerator();
         int[] sizes = { 100_000, 1_000_000, 10_000_000 };
 
         WriteEnvronmentInfo();
